all routing done

+++ src/app/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { User } from "./user.interface";
+import { Order, User } from "./user.interface";
 import { UserModel } from "./user.model";
 
 const createUserIntoDB = async (user: User) => {
@@ -13,7 +13,7 @@ const getAllUsersUserFromDB = async () => {
   return result;
 };
 
-const getSingleUserFromDB = async (userId: string) => {
+const getSingleUserFromDB = async (userId: number) => {
   console.log(userId, 17);
   const existingUser = await UserModel.isExists(userId);
   if (existingUser) {
@@ -24,15 +24,17 @@ const getSingleUserFromDB = async (userId: string) => {
     return existingUser;
   }
 };
-const updateSingleUser = async (userId: string, body: User) => {
+const updateSingleUser = async (userId: number, updatedInfo: User) => {
+  console.log(`Servet hit from update user`);
   const existingUser = await UserModel.isExists(userId);
   if (existingUser) {
+    await UserModel.validate(updatedInfo);
     const result = await UserModel.findOneAndUpdate(
       { userId: userId },
-      { $set: { ...body } },
+      { $set: { ...updatedInfo } },
       { new: true, runValidators: true }
     );
-    console.log(result);
+    console.log(result, 35);
 
     return result;
   } else {
@@ -40,16 +42,96 @@ const updateSingleUser = async (userId: string, body: User) => {
   }
 };
 
-const deleteSingleUserFromDB = async (userId: string) => {
-  const result = await UserModel.findOneAndDelete({ userId });
-  console.log(result, 45);
+const deleteSingleUserFromDB = async (userId: number) => {
+  const existingUser = await UserModel.isExists(userId);
+  if (existingUser) {
+    const result = await UserModel.findOneAndDelete({ userId });
+    console.log(result, 45);
 
-  return result;
+    return result;
+  } else {
+    return existingUser;
+  }
 };
+
+const updateUserOrders = async (userId: number, order: Order) => {
+  const existingUser = await UserModel.isExists(userId);
+  if (existingUser) {
+    const existingOrder = await UserModel.orderExists(userId, order);
+    console.log(existingOrder, 61);
+    if (existingOrder) {
+      return null;
+    } else {
+      const result = await UserModel.updateOne(
+        { userId },
+        { $addToSet: { orders: { ...order } } }
+      );
+      console.log(result, 55);
+
+      return result;
+    }
+  } else {
+    return existingUser;
+  }
+};
+
+const getAllOrdersfromUser = async (userId: number) => {
+  const existingUser = await UserModel.isExists(userId);
+
+  if (existingUser) {
+    const result = await UserModel.find(
+      { userId: userId },
+      { orders: 1, _id: 0 }
+    );
+    console.log(result[0], 86);
+    return result[0];
+  } else {
+    return existingUser;
+  }
+};
+
+const getToalPricefromDB = async (userId: number) => {
+  console.log(`Servet hit from Total Price`);
+  const existingUser = await UserModel.isExists(userId);
+  console.log(existingUser, 96);
+
+  if (existingUser) {
+    const result = await UserModel.aggregate([
+      {
+        $unwind: "$orders",
+      },
+      {
+        $match: { userId: Number(userId) },
+      },
+      {
+        $group: {
+          _id: "$userId",
+          totalPrice: {
+            $sum: { $multiply: ["$orders.price", "$orders.quantity"] },
+          },
+        },
+      },
+
+      {
+        $addFields: { totalPrice: "$totalPrice" },
+      },
+      { $project: { _id: 0 } },
+    ]);
+    console.log(result);
+
+    return result[0];
+  } else {
+    return existingUser;
+  }
+};
+
 export const UserServices = {
   createUserIntoDB,
   getAllUsersUserFromDB,
   getSingleUserFromDB,
   updateSingleUser,
   deleteSingleUserFromDB,
+  updateUserOrders,
+  getAllOrdersfromUser,
+  getToalPricefromDB,
 };
